SERVERLESS-2327 Map environment into a context parameter and mangle HTTP request
This translates our current environment variable based approach into passing a second "context" parameter to the function. The context somewhat resembles Lambda's context parameter but notably keeps the notion of an "activation_id" as that's still an entity we use in our docs and API.

In addition, this implements translation of an HTTP-based event to resemble Lambda's API Gateway event as well. It somewhat closely matches how we do the translation in our Node.js-based Lambda support with the addition of handling the query-string better (which we should probably fix in Node.js as well).