Add support for function signatures - no param and returning nothing.
All of our other runtimes support having a function that has no parameters. They also support having a function that returns nothing. In the case where nothing is returned by the function, the runtime returns an empty dictionary to the rest of the OpenWhisk system. This changes he Python runtime to match that behavior.